Logger: Driverupdate for Windows7 x86 and Windows XP x86, additional registry file, update for compile batches.
git-svn-id: https://intra.ist.tugraz.at/svn/tagstore@348 20732adb-bd90-46d1-8e52-e978120fd635

+++ research_platform/filemonitor/development_stage/flmonflt/FlTracker.c
@@ -736,7 +736,7 @@ POCTRACKER_ENTRY
 
 	while(t->NextEntry)
 	{
-		if (t->Count > last_count)
+		if (t->Count > (unsigned int)last_count)
 		{
 			last_count = t->Count;
 			r = t;
